Refactors code and adds basic token validation.
