+++ constants.py
@@ -9,3 +9,7 @@ PROFILE_KEY = 'profile'
 SECRET_KEY = 'ThisIsTheSecretKey'
 JWT_PAYLOAD = 'jwt_payload'
 
+AUTH0_RESPONSE_TYPE = 'AUTH0_RESPONSE_TYPE'
+AUTH0_GRANT_TYPE = 'AUTH0_GRANT_TYPE'
+AUTH0_SCOPES = 'AUTH0_SCOPES'
+AUTH0_CODE_CHALLENGE_METHOD = 'AUTH0_CODE_CHALLENGE_METHOD'
