+++ dot_env.example
@@ -1,5 +1,9 @@
-AUTH0_DOMAIN=${TENANT_NAME}.auth0.com
 AUTH0_CLIENT_ID=${CLIENT_ID}
 AUTH0_CLIENT_SECRET=${CLIENT_SECRET}
 AUTH0_CALLBACK_URL=http://127.0.0.1:3000/callback
+AUTH0_DOMAIN=${TENANT_NAME}.auth0.com
 AUTH0_AUDIENCE=https://${AUTH0_DOMAIN}/api/v2/
+AUTH0_RESPONSE_TYPE=code
+AUTH0_GRANT_TYPE=authorization_code
+AUTH0_SCOPES=profile openid email read:clients read:rules
+AUTH0_CODE_CHALLENGE_METHOD=S256
